rpc: unify all ledger sequence types to uint32 and stop stringyfying integers < 53-bits wide (#1097)

+++ cmd/soroban-rpc/internal/test/get_ledger_entry_test.go
@@ -107,7 +107,7 @@ func TestGetLedgerEntrySucceeds(t *testing.T) {
 	var result methods.GetLedgerEntryResponse
 	err = client.CallResult(context.Background(), "getLedgerEntry", request, &result)
 	assert.NoError(t, err)
-	assert.Greater(t, result.LatestLedger, int64(0))
+	assert.Greater(t, result.LatestLedger, uint32(0))
 	assert.GreaterOrEqual(t, result.LatestLedger, result.LastModifiedLedger)
 	var entry xdr.LedgerEntryData
 	assert.NoError(t, xdr.SafeUnmarshalBase64(result.XDR, &entry))
